pyproject.toml
[tool.poetry]
name = "simple_test"
version = "0.1.0"
description = ""
authors = ["Konstantin Mamaev <meison@meison.net>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest-splinter = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
